Restructure generic_map into association.rs
